final fix_after. may edit to combine w/ sarah's
this outputs a csv that is useable and has a different column for each of the 10 features of an applicant and each of the 5 features of a property. skylar has to look at sarah's version on collab. she'll probably edit this version to include parts of sarah's because sarah's uses good coding practices, which is something this code may be lacking in.